refactor: move some files around, fix loose ends and fix broken redirect
