+++ redirects-list.json
@@ -1,7 +1,7 @@
 [
   "^https://(?:www\\.)dvc\\.org/(.+)?       https://dvc.org/$1",
   "^https://man\\.dvc\\.org/(.+)?           https://dvc.org/doc/command-reference/$1 303",
-  "^https://error\\.dvc\\.org/(.+)?         https://dvc.org/user-guide/troubleshooting#$1 303",
+  "^https://error\\.dvc\\.org/(.+)?         https://dvc.org/doc/user-guide/troubleshooting#$1 303",
   "^https://(code|data|remote)\\.dvc\\.org/(.+) https://s3-us-east-2.amazonaws.com/dvc-public/$1/$2 303",
   "^/((?:deb|rpm)/.+)                       https://s3-us-east-2.amazonaws.com/dvc-s3-repo/$1 303",
   "^/(?:help|chat)/?$                       https://discordapp.com/invite/dvwXA2N 303",
