+++ server.js
@@ -14,59 +14,56 @@ const { parse } = require('url')
 const next = require('next')
 
 const { getItemByPath } = require('./src/utils/sidebar')
-const { getRedirect } = require('./src/redirects.js')
+const { getRedirect } = require('./src/utils/redirects')
 
 const dev = process.env.NODE_ENV !== 'production'
 const app = next({ dev })
 const handle = app.getRequestHandler()
 const port = process.env.PORT || 3000
 
-// Do not run the server if the module is being required by something else
-if (module.parent == null) {
-  app.prepare().then(() => {
-    createServer((req, res) => {
-      const parsedUrl = parse(req.url, true)
-      const { pathname, query } = parsedUrl
-      const host = req.headers.host
+app.prepare().then(() => {
+  createServer((req, res) => {
+    const parsedUrl = parse(req.url, true)
+    const { pathname, query } = parsedUrl
+    const host = req.headers.host
 
-      /*
-       * HTTP redirects
-       */
-      let [redirectCode, redirectLocation] = getRedirect(host, pathname, {
-        req,
-        dev
-      })
-
-      if (redirectLocation) {
-        // should be getting the query as a string
-        const { query } = parse(req.url)
-        if (query) {
-          redirectLocation += '?' + query
-        }
-        res.writeHead(redirectCode, {
-          'Cache-control': 'no-cache',
-          Location: redirectLocation
-        })
-        res.end()
-      } else if (/^\/doc(\/.*)?$/.test(pathname)) {
-        /*
-         * Docs Engine handler
-         */
-
-        // Force 404 response code for any inexistent /doc item.
-        if (!getItemByPath(pathname)) {
-          res.statusCode = 404
-        }
-
-        // Custom route for all docs
-        app.render(req, res, '/doc', query)
-      } else {
-        // Regular Next.js handler
-        handle(req, res, parsedUrl)
-      }
-    }).listen(port, err => {
-      if (err) throw err
-      console.info(`> Ready on localhost:${port}`)
+    /*
+     * HTTP redirects
+     */
+    let [redirectCode, redirectLocation] = getRedirect(host, pathname, {
+      req,
+      dev
     })
+
+    if (redirectLocation) {
+      // should be getting the query as a string
+      const { query } = parse(req.url)
+      if (query) {
+        redirectLocation += '?' + query
+      }
+      res.writeHead(redirectCode, {
+        'Cache-control': 'no-cache',
+        Location: redirectLocation
+      })
+      res.end()
+    } else if (/^\/doc(\/.*)?$/.test(pathname)) {
+      /*
+       * Docs Engine handler
+       */
+
+      // Force 404 response code for any inexistent /doc item.
+      if (!getItemByPath(pathname)) {
+        res.statusCode = 404
+      }
+
+      // Custom route for all docs
+      app.render(req, res, '/doc', query)
+    } else {
+      // Regular Next.js handler
+      handle(req, res, parsedUrl)
+    }
+  }).listen(port, err => {
+    if (err) throw err
+    console.info(`> Ready on localhost:${port}`)
   })
-}
+})
